Introduced java interface named GameConsole
By introducing Java Interface, in GameRunner class, we mention the common interface instead specifying the each. so that in future if we want to change the game.. we just need to modify the main class.

this is one of the solution to tightly coupled java code issue. (Loose coupling using Java Interface)